Fixed "table AS" incompatibility with Oracle databases
Oracle database does not support the use of "AS" in table aliases in queries.
For instance:
SELECT foo
FROM bar AS baz

yields the following error in Oracle:
ORA-00933: SQL command not properly ended

The solution is to omit the "AS". This was done in several locations affecting
the following functionality:
- "Unlink Users" administration link
- "Files" administration link
- "Submission Inbox" in-assignment tab